refatoracao do codigo do back-end para recuperar dado do usuario pelo jwt ao inves do cpf, e utilizar padroes recentes de injecao de dependecia e lombok recomendados pelo spring

+++ back-end/src/main/java/com/petcemetery/petcemetery/DTO/AquisicaoJazigoDTO.java
@@ -2,13 +2,15 @@ package com.petcemetery.petcemetery.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AquisicaoJazigoDTO {
 
     String endereco;
-    double valor;
+    BigDecimal valor;
 
-    public AquisicaoJazigoDTO(String endereco, double valor) {
+    public AquisicaoJazigoDTO(String endereco, BigDecimal valor) {
         this.endereco = endereco;
         this.valor = valor;
     }
