+++ back-end/src/main/java/com/petcemetery/petcemetery/DTO/CarrinhoDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,5 +14,5 @@ public class CarrinhoDTO {
     String jazigoId;
     String selectedOrnament;
     String tipo;
-    double valor;
+    BigDecimal valor;
 }
