+++ back-end/src/main/java/com/petcemetery/petcemetery/DTO/ContratoDTO.java
@@ -4,19 +4,21 @@ import com.petcemetery.petcemetery.model.Servico.ServicoEnum;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContratoDTO {
 
     String cpfCliente;
-    long idJazigo; // Pode ser nulo!
-    double valor;
+    Long idJazigo; // Pode ser nulo!
+    BigDecimal valor;
     ServicoEnum tipoServico;
     String enderecoJazigo; // Pode ser nulo!
-    long idPet; // Pode ser nulo!
+    Long idPet; // Pode ser nulo!
     String dataServico;
 
     // CONSTRUTOR - Full args
-    public ContratoDTO(double valor, ServicoEnum tipoServico, String enderecoJazigo, long idJazigo, long idPet, String dataServico, String cpfCliente) {
+    public ContratoDTO(BigDecimal valor, ServicoEnum tipoServico, String enderecoJazigo, Long idJazigo, Long idPet, String dataServico, String cpfCliente) {
         this.idJazigo = idJazigo;
         this.valor = valor;
         this.tipoServico = tipoServico;
@@ -27,7 +29,7 @@ public class ContratoDTO {
     }
 
     // CONSTRUTOR - Pet nulo
-    public ContratoDTO(double valor, ServicoEnum tipoServico, String enderecoJazigo, long idJazigo, String cpfCliente) {
+    public ContratoDTO(BigDecimal valor, ServicoEnum tipoServico, String enderecoJazigo, Long idJazigo, String cpfCliente) {
         this.idJazigo = idJazigo;
         this.valor = valor;
         this.tipoServico = tipoServico;
@@ -36,7 +38,7 @@ public class ContratoDTO {
     }
 
     // CONSTRUTOR - Pet e Jazigo nulos
-    public ContratoDTO(double valor, ServicoEnum tipoServico, String cpfCliente) {
+    public ContratoDTO(BigDecimal valor, ServicoEnum tipoServico, String cpfCliente) {
         this.valor = valor;
         this.tipoServico = tipoServico;
         this.cpfCliente = cpfCliente;
